feat: dto 구조 변경 및 필드 추가
- processId로 사용자 구분

#32
